+++ src/netcdf/input_netcdf.c
@@ -363,7 +363,7 @@ Input_netcdf openinput_netcdf(const char *filename, /**< filename */
   }
   input->lon_min=dim[0];
   if(input->lon_len==1)
-    input->lon_res=config->resolution.lon;
+    input->lon_res=(float)config->resolution.lon;
   else
     input->lon_res=(dim[input->lon_len-1]-dim[0])/(input->lon_len-1);
   if(fabs(input->lon_res-config->resolution.lon)/config->resolution.lon>1e-3)
@@ -412,7 +412,7 @@ Input_netcdf openinput_netcdf(const char *filename, /**< filename */
   if(input->lat_len==1)
   {
     input->lat_min=dim[0];
-    input->lat_res=config->resolution.lat;
+    input->lat_res=(float)config->resolution.lat;
     input->offset=0;
   }
   else if(dim[1]>dim[0])
